Correções de Controller Books

+++ WebApiRestUdemy/Migrations/20220112145336_Initial.cs
@@ -15,7 +15,7 @@ namespace WebApiRestUdemy.Migrations
                 name: "books",
                 columns: table => new
                 {
-                    id = table.Column<int>(type: "int", nullable: false)
+                    id = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("MySql:ValueGenerationStrategy", MySqlValueGenerationStrategy.IdentityColumn),
                     author = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
